middleware: extract pool-assignment state and draw logic to scheduling/pool_assignment.py (#152 step 5)
19 methods moved verbatim to module functions taking the builder as first
parameter; ScheduleWorkbookBuilder keeps staticmethod aliases and thin
wrappers so church_teams_export.py's __dict__ delegation and existing
tests keep working. No behavior changes.
